handle change user info

+++ server/controllers/userController.js
@@ -14,10 +14,12 @@ export const getUserById = async (req, res) => {
   }
 };
 
+// Cập nhật thông tin người dùng (admin)
 export const updateUser = async (req, res) => {
   const { id } = req.params;
   const { fullName, email, avatar, role } = req.body;
 
+  // Kiểm tra quyền truy cập
   if (req.user.role !== "admin") {
     return res.status(403).json({ message: "Access denied" });
   }
@@ -38,6 +40,31 @@ export const updateUser = async (req, res) => {
   }
 };
 
+// Cập nhật thông tin người dùng (all user)
+export const updateUserInfo = async (req, res) => {
+  const { id } = req.params;
+  const { fullName, email } = req.body;
+
+  // Kiểm tra quyền truy cập
+  if (!req.user) {
+    return res.status(403).json({ message: "Access denied" });
+  }
+
+  try {
+    const user = await User.findByIdAndUpdate(
+      id,
+      { fullName, email },
+      { new: true, runValidators: true }
+    ).select("-password");
+
+    if (!user) return res.status(404).json({ message: "User not found" });
+
+    res.status(200).json(user);
+  } catch (error) {
+    console.error("Error updating user info:", error);
+    res.status(500).json({ message: "Something went wrong" });
+  }
+};
 export const getAllUsers = async (req, res) => {
   try {
     const users = await User.find().select("-password");
@@ -47,8 +74,15 @@ export const getAllUsers = async (req, res) => {
     res.status(500).json({ message: "Something went wrong" });
   }
 };
+
+// Xóa người dùng theo ID
 export const deleteUser = async (req, res) => {
   const { id } = req.params;
+
+  // Kiểm tra quyền truy cập
+  if (req.user.role !== "admin") {
+    return res.status(403).json({ message: "Access denied" });
+  }
 
   try {
     const user = await User.findByIdAndDelete(id);
